+++ bindings/python/pymongocrypt/asynchronous/explicit_encrypter.py
@@ -93,7 +93,7 @@ class AsyncExplicitEncrypter:
         contention_factor=None,
         range_opts=None,
         is_expression=False,
-        text_opts=None,
+        string_opts=None,
     ):
         """Encrypts a BSON value.
 
@@ -115,7 +115,7 @@ class AsyncExplicitEncrypter:
             with the "range" algorithm encoded as a BSON document.
           - `is_expression` (boolean): True if this is an encryptExpression()
             context. Defaults to False.
-          - `text_opts` (bytes): Options for explicit encryption
+          - `string_opts` (bytes): Options for explicit encryption
             with the "string" algorithm encoded as a BSON document.
 
         :Returns:
@@ -127,6 +127,8 @@ class AsyncExplicitEncrypter:
            Added the `range_opts` and `is_expression` parameters.
         .. versionchanged:: 1.16
            Added the `text_opts` parameter.
+        .. versionchanged:: 1.19
+           Renamed the `text_opts` parameter to `string_opts`.
         """
         # CDRIVER-3275 key_alt_name needs to be wrapped in a bson document.
         if key_alt_name is not None:
@@ -139,7 +141,7 @@ class AsyncExplicitEncrypter:
             contention_factor,
             range_opts,
             is_expression,
-            text_opts,
+            string_opts,
         )
         with self.mongocrypt.explicit_encryption_context(value, opts) as ctx:
             return await run_state_machine(ctx, self.callback)