+++ bindings/python/pymongocrypt/options.py
@@ -170,7 +170,7 @@ class ExplicitEncryptOpts:
         contention_factor=None,
         range_opts=None,
         is_expression=False,
-        text_opts=None,
+        string_opts=None,
     ):
         """Options for explicit encryption.
 
@@ -187,7 +187,7 @@ class ExplicitEncryptOpts:
             with the "range" algorithm encoded as a BSON document.
           - `is_expression` (boolean): True if this is an encryptExpression()
             context. Defaults to False.
-          - `text_opts` (bytes): Options for explicit encryption
+          - `string_opts` (bytes): Options for explicit encryption
             with the "string" algorithm encoded as a BSON document.
 
         .. versionchanged:: 1.3
@@ -196,6 +196,8 @@ class ExplicitEncryptOpts:
            Added the `range_opts` and `is_expression` parameters.
         .. versionchanged:: 1.16
            Added the `text_opts` parameter.
+        .. versionchanged:: 1.19
+           Renamed the `text_opts` parameter to `string_opts`.
         """
         self.algorithm = algorithm
         self.key_id = key_id
@@ -217,11 +219,11 @@ class ExplicitEncryptOpts:
             )
         self.range_opts = range_opts
         self.is_expression = is_expression
-        if text_opts is not None and not isinstance(text_opts, bytes):
+        if string_opts is not None and not isinstance(string_opts, bytes):
             raise TypeError(
-                f"text_opts must be an bytes or None, not: {type(text_opts)}"
+                f"string_opts must be an bytes or None, not: {type(string_opts)}"
             )
-        self.text_opts = text_opts
+        self.string_opts = string_opts
 
 
 class DataKeyOpts: